Set image dimensions as int32

+++ pkg/module/gallery/api/v1/api_v1_image.go
@@ -17,8 +17,8 @@ type ImageV1 struct {
 	FileName     string                `json:"fileName"`
 	FileSize     int64                 `json:"fileSize"`
 	MimeType     string                `json:"fileType"`
-	Width        int16                 `json:"width"`
-	Height       int16                 `json:"height"`
+	Width        int32                 `json:"width"`
+	Height       int32                 `json:"height"`
 }
 
 type ImageTranslationV1 struct {
